requirements.txt
psutil
wmi
python-dotenv
requests
